+++ src/utils.ts
@@ -190,6 +190,52 @@ export async function resetCache() {
   Logger.log("[vscode-sfdx-hardis] Reset cache");
 }
 
+export async function execCommandWithProgress(
+  command: string,
+  options: ExecCommandOptions = {
+    fail: false,
+    output: false,
+    debug: false,
+    spinner: true,
+  },
+  progressMessage: string,
+) {
+  return await vscode.window.withProgress(
+    {
+      location: vscode.ProgressLocation.Notification,
+      title: progressMessage || "Executing command...",
+      cancellable: false,
+    },
+    async () => {
+      return await execCommand(command, options);
+    },
+  );
+}
+
+/* jscpd:ignore-start */
+export async function execSfdxJsonWithProgress(
+  command: string,
+  options: ExecCommandOptions = {
+    fail: false,
+    output: false,
+    debug: false,
+    spinner: true,
+  },
+  progressMessage: string,
+) {
+  return await vscode.window.withProgress(
+    {
+      location: vscode.ProgressLocation.Notification,
+      title: progressMessage || "Executing command...",
+      cancellable: false,
+    },
+    async () => {
+      return await execSfdxJson(command, options);
+    },
+  );
+}
+/* jscpd:ignore-end */
+
 // Execute command
 export async function execCommand(
   command: string,
@@ -473,7 +519,7 @@ export async function loadExternalSfdxHardisConfiguration() {
 }
 
 // Fetch remote config file
-/* jscpd-ignore-start */
+/* jscpd:ignore-start */
 async function loadFromRemoteConfigFile(url: string) {
   if (REMOTE_CONFIGS[url]) {
     return REMOTE_CONFIGS[url];
@@ -491,7 +537,7 @@ async function loadFromRemoteConfigFile(url: string) {
   REMOTE_CONFIGS[url] = remoteConfig;
   return remoteConfig;
 }
-/* jscpd-ignore-end */
+/* jscpd:ignore-end */
 
 export async function readSfdxHardisConfig(): Promise<any> {
   if (vscode.workspace.workspaceFolders) {